Add comments (filtered) and categories
The filter for comments is simple: Remove comments that contain the
string 'http' and the author has 0 submissions to the journals.
This effectively has wiped out all the spam, with only a few false positives.

The publication metadata is now wrapped inside a `publication: {}` key,
following changes introduced in #22.
However this changes:
`publication: {id:..}` to `publication: {publication_id}`

Closes #17
Closes #19