Add citation_list to metadata
With the format
```
"revisions": [
    {
    "citation_list": [
        {
            "doi": "10.1023/b:visi.0000029664.99615.94",
            "key": "ref1",
            "score": 84.210754,
            "unstructured": "Distinctive image features from scale-invariant keypoints+IJCV+60+2+91+110+November 2004+D. G. Lowe"
        },
    ]
    }
]
```

`doi`: If a publication has no doi, it's because the score is slower than 60
why 60? Chosen by trial and error as a good tradeoff. It is high enough
to not let false positive matches, and low enough for some publication
with not much metadata, but with title matching perfectly etc.

`key`: The key parsed by CERMINE

`unstructured`: The plain text used as query in the crossref API as `query.bibliographic`
CERMINE gives the output as a structured XML, but  it is not 100%
reliable, we just concatenate all the fields into a plain text query
here.

`score`: The score of the FIRST match from crossref using the value of `unstructured` as the query.
if the match is lower than 60, we don't populate a doi to that reference.

These are the publications that were not updated because they don't have
a pdf, or a pdf without references. These are probably not errors, just
odd publications, data, workshops, etc.

```
No citations_crossref.json file found for pub 105, skipping:

105, 113, 114, 117, 126, 128, 129, 138, 140, 141, 143, 148, 150, 152, 153, 160, 168, 178, 181, 186, 187, 189, 191, 196, 200, 201, 208, 188, 285, 54, 216, 240, 242, 244, 247, 249, 272, 305, 307, 318, 319, 323, 42, 45, 46, 543, 544, 55, 553, 56, 577, 58, 580, 59, 591, 61, 62, 629, 63, 64, 646, 652, 664, 67, 671, 676, 680, 689, 693, 695, 696, 698, 702, 704, 705, 707, 708, 709, 711, 712, 716, 721, 724, 725, 726, 74, 741, 742, 750, 751, 764, 767, 774, 78, 782, 783, 785, 789, 792, 797, 801, 802, 81, 821, 832, 835, 836, 84, 840, 844, 85, 858, 86, 870, 879, 88, 890, 894, 915, 918, 920, 921, 93, 950, 969, 973, 981, 990, 992
```

Read the discussion in #70 for the workflow used to extract this information.
The scripts used to perform this merge are in this repo:
https://github.com/phcerdan/insightjournal-dev/commit/0fe4c98d21d3ac69f8d60d59e90ea9d71ed6d8f5

Closes #70