Add abstract to crossref submission
Abstract has to be formatted in jats format.
See https://www.crossref.org/documentation/content-registration/descriptive-metadata/abstracts/
Basically jats allow math formulas in a really complex xml format.

We use pandoc-jats (need to be installed in the system) to read our
stored abstract in metadata.json and transform them to jats.
We use a thin wrapper of pandoc in python: pypandoc, to use it.

Tested with local schema validation, and uploading to test server in
crossref.